Change build to create a static site for an SPA.
After some struggles trying to see if we could deploy a node server
with cloud functions for cluster to enable SSR, we gave up on that
and went back to serving the way we've done other apps like TODO: as
a single page application served as a statically compiled bundle of
html/css/javscript. This model should work well for cluster because
most if not all of the pages will need to load game code to render,
and if we really want SSR for something we can turn it on for specific
routes that are carefully coded to work with SSR.

The specific roadblock we'd hit when trying to get SSR to work has
something to do with redux being a "CommonJS" module and not importing
correctly in the node environment.