'patch' updates version number & git & deploy.
First off, use a rollup plugin so that bin/version gets
executed on every build and creates a version.ts file that
embeds the version string for the app in src/lib. Then the
code is free to display the version string wherever it wants
so that we can tell what version is deployed - for now, we
just show this at the bottom of the page. The version number
comes from package.json.

'bun version patch' updates the version number in package.json,
and deploys the new code to firebase. If that succeeds, it then
runs bin/packagedist which checks it in with a new git tag, and
pushes it up to github.