First approach to signed in / signed out state.
Our experience with TODO was that having a different group of
pages just for signed in vs signed out slowed down load time
needlessly with redirects to get to the page the user is on,
and we never quite figured out the cleanest way to fix that.

So for Cluster, where we want to have some (potentially large) set of
pages that the user can visit before they decide to commit to the game
by signing in, we'll use a (website) group to enable serving of the
signed-out experience, with help, examples, screenshots, and so on.

However to really do anything the user has to sign in, and the
(game) group of pages will require that they are signed in right
at the top level in +layout.svelte, and force the user through a
sign-in process before they can see their game or the game lobby.

Then the user experience should be if they visit a deep link like
https://cluster.com/game?id=1234 and they are signed out, they first
sign in and then are immediately served game #1234. If they visit
the top level pages they see the website under (website). Normally
speaking, the user will be able to go directly to the lobby or to
a game as they'll already be signed in and can access those pages
without anything more than an internal check of their signed in
state (i.e. no redirecting).