RealtimeDB and Cloud functions.
Turn on the realtime database and write to it from the profile page.
We need to make a file for the permissions rules and check it in.

Turn on cloud functions and monitor the realtime database for changes to
the profiles of all users.  For now, print a log message containing the
new value to cloud logging.

Biome is unaware of svelte, so it complains about several svelte-isms,
like when a variable is used in a 'bind', and '$:'.