Full-circle data: browser->realtime db->firestore->browser
A change in a web page (the user's profile name) is sent to the realtime
database for storage, which triggers a cloud function that stores the
profile information in firestore.

The web page is subscribed to the profile information, and when the data
is modified in firestore, the page gets a notification and updates its
view of all the profiles.