First example of how we'll share code between client/server.
We talked about either having three sibling projects that depend
on each other npm-style, or about having common code live in the
functions project which is a sort of "fake child" of the top level
client project by containment.

As far as we know the rules of the game are that the functions code
can't reach directories above its package.json, but the client code
can freely reach into the functions directory to get at the common
code.

In an effort to make this somewhat generic to future changes, we
configured a $common path in tsconfig.json for both the functions
source and the client source so that imports from $common/<module>
will refer to the client project's functions subdirectory, namely
functions/src/common/<module>, and this enables importing of common
code and types in the same way from both the client code and the
server side code. If we later move it, we'll need to redo the path
definition for $common in the two tsconfig.json files.