First cut of joinGame/leaveGame actions.
Add a lobby page which shows available games to the user,
and lets them join or leave a game.

Spend a bunch of time thinking about the overall design, in which
we decided game actions would accumulate as pending actions in the
realtime DB, and each client could apply pending game actions using
shared reducers/action executing code to their current gamestate;
while when the tick happens the server would wake up, mark actions
as executed, and generate a new server side game state for everyone
to observe.

When a client gets a new game state (either from a first time load
or from a new tick happening) it must then query the game actions
for as yet unexecuted actions, execute those on top of the game state
and present the UI to the user.

We convinced ourselves that sharing actions could work by copying them
to a location where the other user could read them, and that if we
implement that it will be on at least an entire tick basis (i.e. no
sharing of data that only covers part of what you did in tick N).

Unsolved issue: we couldn't use patch from @ourway/patch, so we just
copied it into the functions code for now.