Factor out the game reducer into a standalone function.
Take all the database logic and centralize it in a single
executeGameAction function that relies on a game reducer to
actually mutate the game state appropriately.

This should form the basis for pulling the game reducer out
into a common/game.ts file that can be used on both the client
and server side and tested in isolation from the database code.