Much futzing around with versions to pacify bun.
We'd gotten into a bad state where you could either run

bun i -D c8

and then use bun run test:watch to watch tests and create coverage,
but you couldn't run 'bun build', so you had to rm  the c8 package
to build.

Obviously we want to be able to run tests _and_ build and the
secret incantation that seemed to get us out of npm version hell
was to run

bun i -D @vitest/v8-coverage

instead of installing c8 directly, and this was able to install
the conflicting packages in such a way that everything is happy
and running tests works as well as building.

The output of tests as shown in neovim is a little odd but usable,
and the code coverage output changed in ways that aren't wrong but
weren't deliberate, but ... this "mostly works".