Changes to how we run firebase-tools.
The firebase-tools package directions tell you to install the tools
globally with a command like

npm install -g firebase-tools@latest

but this means we have no real control over what version of
firebase-tools is being used in various projects. It turns out that
since the 'firebase' command doesn't have the same name as the
package, using npx or bunx on it doesn't work unless the tools are
installed locally; which is great because this happy accident means
that

bunx firebase login

will fail if run in a directory where the firebase-tools package
can't be found locally, so we can now be assurred that the version
of firebase that we're using is the one installed for the particular
project we're working on and not whatever happens to be globally
installed. In fact we were forced into this by a version mismatch
between firebase-functions (which is installed locally) and
firebase-tools (which previously was installed globally) creating
a mysterious error message about not being able to parse the build
configuration and therefore being unable to load our firebase
functions into the emulator.

Meanwhile, we also discovered that it is reasonably easy to be
logged into multiple accounts using firebase-tools. We didn't manage
to figure out where it stores this state, but you can run:

bunx firebase login:list

to see your logins, login:add to add a login, and login:use to set
the account you want for the project you are in. So this is easier
than constantly having to use bunx firebase logout/login to switch
accounts every time you switch projects though the lack of understanding
of where the magic is is slightly unsatisfying.

Finally, we were frustrated by the fact that tsc's default behaviour
(and esbuild's) is to not rewrite paths that have aliases in
tsconfig.json. This doesn't matter at all if you only import types
from say $common/gamestate.ts, but when you import functions that
turns into a require("$common/gamestate.js") in the generated code
and then node can't find the file because $common doesn't exist as
a path (and symlinking it also doesn't work). For now, suffer with
this and use import of the true path name if more than types are
imported, and use the $common alias when it is just types that are
being imported. There is something called a 'resolver' or similar
in esbuild that can be overridden to change this behaviour, but we
were frustrated enough to just call it a day in an effort to get
to where we can actually do something.