Finish writing the test for leave game.
While trying to figure out why a 'null' patch was being written
when andrew joined and then left a game, we saw that the leave
game test wasn't actually expecting anything. Add expect calls
so that this test actually validates that the reducer works.

Having proven that the reducer worked, had to go fix the bug
elsewhere...