Fix leaving games to work properly.
The actual problem was that the initial game state was never written
as a patch to the first block, and that was because the diff would
see it as always being there as a result of it getting set up in
getGameState(). Change getGameState() to write this initial patch
so that the rest of the definitionally "bug-free" code would work
(the database code lacks any tests).