A hacky first cut of getting games to load.
Write a first "Open" button that takes the user to the
game URL .../game?id=<gameid> and on that page, load and
replay the game patches to show the game state.

When a new player joins or leaves the required patch is
applied from where we left off, and overall this completes
a proof of concept of the whole rube goldberg machine:

1. user takes a UI action and it is written to realtime db
2. server side function sees the change and updates the game, generating
   a patch for the new game state and writing it to the appropriate
   block in the game's gamestate blocks
3. the client sees the gamestate block update if the game is loaded,
   reloads all the patches for that block skipping down to the first one
   whose timestamp is newer than the last thing seen, and replays the
   patches from there

So now we can actually start to implement server side modifications
of the game state and see the update in real time on the client. The
client will never modify the game state directly, but will trigger
modifications by writing requests into the realtime database that cause
the server to take action.

TODO: once again, we couldn't import patch properly. We will need to
spend a session figuring out why our package is difficult to import
and fixing it :(