Make @ourway patch work by bundling.
Andrew was playing with @ourway/patch and we weren't seeing any
problems until we looked at the functions emulator, and realized
that it can't figure out how to import the patch.ts file from the
@ourway/patch package.

Now it isn't clear if we're supposed to compile up the package
differently to be nice package maintainers who produce a package
that can be used in node, typescript, and browser javascript
environments, or if it's OK to just bundle the code in the client
codebase and say "this is a typescript-only project, you deal with
it if you want something else".

Nonetheless the way we went was to use esbuild in the functions
project to build the development version and this works just fine,
producing a single bundled 7.5Mb file with all the source in it for
the server side.

We did not fix the `bun build` script, so when we come to the point
where we want to deploy the functions for real this will bite us
again and make us reconfirm the decision or change how we package
@ourway/patch and other libraries.