Get game creation and starting working.
On the server side, implement a tick() method meant to be called
every tick interval, that loops through all the games that have
been created and either tells them to compute their tick, or starts
them if they're not yet started. As part of this, added a bunch of
new metadata to the options, namely a started boolean and a
playersNeeded count so that we don't have to instantiate all the
game states for this hopefully easy to keep track of metadata.

Then the server starts the next instance of the N player game if
autospawn is on for the game.

On the client side, show the new metadata. In the lobby, show a
disabled "full" button if the game is ready but not started, and
in the game screen itself display the started state and tick # so
that we can see those things change as we call tick() from the
functions shell.

TODO: we know the server side code which updates multiple firebase
objects sequentially is bad. We will have to (at some point in the
far future) look into how to avoid the race conditions,  presumably
with some sort of 'firebase transaction' that we can wrap around
all the db operations to make them atomic.