Complete games.
A game is completed if it has a winner.  Look in GameOptions.started to see if
a game is waiting to start or has started, then look in GameOptions.winner to
see if the game is finished.

Randomly end a game and randomly choose a winner.  And modify the UI to show a
winner, and also if players are needed or if there are enough players but we're
waiting for the first tick to happen.

We moved logic that modified the game state into the game() reducer.  This
includes modifying playersNeeded for both join_game and leave_game, and
modifying started in start_game.  As part of this change, executeGameAction
notices if anything has changed in nextstate.options and automatically udates
the firebase doc for the game at /games/gameid.  This makes it easier to write
reducer code without forgetting to update the doc.