Move players list from main game state to options.
We seem to be converging on using GameOptions for all game metadata
that we want easy access to, so move the players list out of the
gamestate and into its options.

This will enable a follow-on change to get rid of the hacky recording
of what games a player has joined in their profile, since the
metadata are now readily available from the games directly.

While we were doing this we discovered that saving an empty object
into realtime database doesn't work - the key and value just vanish
- so though we write the empty players list into the "create_game"
action in realtime db, it gets lost. So in a somewhat unsatisfying
way, make sure that createGame() in the functions source code always
writes the empty players object into the options - which does work
as expected in firestore.

While we're at it, notice that 'untrusted' options like the number
of players needed or who is in the game have to be passed into
createGame even though this is needless/pointless; maybe game
metadata needs to be split into those options we think a user should
be able to set and those pieces of metadata that ought to be computed
or known automatically and not subject to what the user says.