Remove special cases for join_game and leave_game.
Now that the reducer for the game updates the player list in the
game metadata, we don't have to hack it into the individual player's
profiles anymore. Change the UI to get the info from the game
document directly, and remove the special case code from the server
side that added this extra info specially for join_game and leave_game
actions.

In the process we noticed that the load order of the lobby is a little
odd, in that it's still important to update the game list after the user
profiles are loaded, but we shrugged that off for now as it is all
working.