Add type checking to tsc.
Type checking is enabled with the --typecheck option to vitest.  The
typecheck option in vitest.config.ts looks for and run tests in files
with names like *.test-d.ts.

The type-checking in type-tests is done at compile time, and not at test
time.  A test function with expectTypeOf() can be failing, but this
won't be detected during the test run.

I have purposely left a tsc compile error in profiles.test-d.ts
until we have a way to stop us from checking in compile errors.

The above changes forced a change to the husky pre-commit script because
it was doing 'bun test' which was failing.  'bun test' finds all test
files including the generated ones in lib/*.test.js, which couldn't find
their imports from 'common'.  Change the husky pre-commit script from
'bun test' to 'bun run test' which runs the tests the way they are done
from package.json.

Also, repeat the pre-commit steps for the 'functions' sub-project too.