First cut of map generation.
Pull over the map prototype code that we wrote and jam it into the
"real game".

This version makes an encrypted set of stars each with their own
encryption key, and gives every player who joins the game the ability
to decrypt the whole galaxy (for now). We have flexibility to
restrict what users can see when they first join as a result of how
the keys are set up; the player will ignore all those stars they
can't see, and all they can really determine is how big the galaxy
is before the game starts if we hide some of the stars in the future.

Then, in the UI, added a map widget that shows the galaxy rendered
in 3d and lets you rotate it and zoom it and select stars to lay
out "routes".