Change star coloration to be false colour.
The color based on the real star temperature just doesn't offer
enough video-game UI options; it ranges from red to white to white
to whiter and there are only a few colours that look OK or are even
distinguishable.

Change the shader to take a low and high color of any value and
interpolate them to create a reasonable false-colour star effect.

We can now use the star colour either to represent player ownership
or star type using colours that are perceptibly different, as well
as colour the stars for selection/hover/whatever we might want for
the UI.

In the process we spent several hours debugging problems where the
shader just stops working if you set any value differently in the
uniforms object. For any value that's changing, it seems important
to initialize it to a _constant_ first, and then override that with
an assigment to that specific key of the uniforms object. We didn't
really understand _why_ this is so but it took several hours to
figure out that it's important that the initializer for the uniforms
is all constants.