First cut of generating and displaying star names.
Bring over Andrew's star names code from the other branch, and add
prefixes and suffixes to reduce the chance of collisions.

Then, use the <HTML> extra to get star names on the screen, and
fight with understanding coordinate systems to adjust correctly.
Every object in three.js (and therefore threlte) has its own
coordinate system for defining its geometry, including the camera
whose local coordinate system has the Z axis pointed into the center
of the screen. Capitalize on this by moving the position for the
label into camera space, adjusting it in Y (which is for all intents
and purposes the same as "down" in screen space) and then transforming
it back, pinning the text underneath the center of the star it goes
with.

We still need to account for how zoomed in our out the user is and
adjust the font size (or perhaps suppress the name altogether) but
that will be a task for next week.