Show fake data to get star info the right size.
Modify the star display to show name, infrastructure, and ships,
so that we can experiment with scaling/hiding the per-star info.

Scale it based on distance to the star, or hide it if you're far
from the target star, but always show it if you hover over a particular
star.

Also tone down the gray colour of the projection to the galactic plane.
Tried the display wihtout it altogether but the projection seems to
really help orient when you're looking at the galaxy.

All this could be tuned to be nicer but seemed good enough to be getting
on with to try to get closer to an actual game to see how the experience
is in practice.

Also we made a note of what the user wants to do / why we're bothering
to display the info at all:

- names for flavour
- infrastructure to determine if the star is valuable/worth capturing
- strength to know how well defended the star is
- type of star to know affinity to your faction
- ownership to know if the star is friend or foe

Star type is going to be represented by colour, in an upcoming change.

Ownership maybe will be represented by an icon or avatar near the star,
as yet undecided.