First cut of homestar ownership.
Update all stars to have an "owner" field which is either "nobody" or
the uuid of the star's owner.

When a player joins a game, find an available homestar and assign it to
them. Later, this should also set the star type to the type that this
faction has an affinity for, but since we don't have factions, faction
affinities, or costs for anything yet ... it can wait.

Hack the map drawing code to display hte owner UID below the star and
make homestars huge so that they're easy to spot. Both short term hacks
that must be removed.

Finally, change the way we process patches to only re-decrypt objects
that have changed, or that have never before been decrypted. That way
when galaxy updates come in as players join the game, only one star gets
processed instead of re-decrypting the entire galaxy needlessly.