Trying to add hardcoded developments.
We decided the path of least resistance to being able to route a
fleet with infantry and fighters in it from place to place was to
give each homestar a bit of production by default. So hardcode the
homestar assignment to give each star 0/1/1/0 by default, and plumb
that through to the UI code where it is helpfully console.log'ed and
ignored.

Meanwhile, it seemed like it'd be better to pass maps instead of arrays
from the UI code into the galaxy screen, and got that working but it is
unfinished because we still haven't renamed objects and removed the
wasteful array construction from the calling code.

So next steps would be to clean that up and actually display the correct
developments information under each star; and then add actual production
to the tick code!